Replace config/data model with common modules in purchase dir
