+++ backend/compact-connect/lambdas/common-python/cc_common/data_model/schema/jurisdiction.py
@@ -1,13 +1,13 @@
 # ruff: noqa: N801, N815, ARG002 invalid-name unused-kwargs
 from collections import UserDict
 
-from config import config
+from cc_common.config import config
 from marshmallow import Schema, pre_dump
 from marshmallow.fields import Boolean, Decimal, List, Nested, String
 from marshmallow.validate import Length, OneOf
 
-from data_model.schema.base_record import BaseRecordSchema, ForgivingSchema
-from data_model.schema.common import CCEnum
+from cc_common.data_model.schema.base_record import BaseRecordSchema, ForgivingSchema
+from cc_common.data_model.schema.common import CCEnum
 
 JURISDICTION_TYPE = 'jurisdiction'
