--- backend/compact-connect/lambdas/purchases/data_model/schema/__init__.py
@@ -1,7 +1,0 @@
-# ruff: noqa: F401
-# We import all the record types with the package to ensure they are all registered
-from .compact import CompactRecordSchema
-from .jurisdiction import JurisdictionRecordSchema
-from .license import LicenseRecordSchema
-from .privilege import PrivilegeRecordSchema
-from .provider import ProviderRecordSchema
