+++ backend/compact-connect/lambdas/purchases/handlers/utils.py
@@ -8,7 +8,7 @@ from uuid import UUID
 
 from aws_lambda_powertools.utilities.typing import LambdaContext
 from botocore.exceptions import ClientError
-from config import logger
+from cc_common.config import logger
 from cc_common.exceptions import CCAccessDeniedException, CCInvalidRequestException, CCNotFoundException, CCUnauthorizedException
